fix: address CodeRabbit review feedback on SEO and loaders
