+++ .gitignore
@@ -63,6 +63,6 @@ pip-delete-this-directory.txt
 .python-version
 
 # Generated SEO / IndexNow Files
-public/indexnow-key.txt
-public/*indexnow*.txt
+public/*.txt
+!public/robots.txt
