+++ scripts/generate-indexnow.js
@@ -1,46 +1,15 @@
 const fs = require('fs');
 const path = require('path');
-
-function loadEnv() {
-  const env = { ...process.env };
-  const envPaths = [
-    path.join(__dirname, '../.env.local'),
-    path.join(__dirname, '../.env')
-  ];
-  for (const envPath of envPaths) {
-    if (fs.existsSync(envPath)) {
-      try {
-        const content = fs.readFileSync(envPath, 'utf8');
-        content.split('\n').forEach(line => {
-          const match = line.match(/^\s*([\w.-]+)\s*=\s*(.*)?\s*$/);
-          if (match) {
-            const key = match[1];
-            let value = match[2] || '';
-            if (value.startsWith('"') && value.endsWith('"')) {
-              value = value.slice(1, -1);
-            } else if (value.startsWith("'") && value.endsWith("'")) {
-              value = value.slice(1, -1);
-            }
-            env[key] = value.trim();
-          }
-        });
-      } catch (err) {
-        // Silent catch
-      }
-    }
-  }
-  return env;
-}
+const { loadEnv } = require('./load-env');
 
 const env = loadEnv();
-// Fallback to a stable default key if VITE_INDEXNOW_KEY is not defined
-const INDEXNOW_KEY = env.VITE_INDEXNOW_KEY || 'moviemonkai123indexnowkey';
+const INDEXNOW_KEY = env.VITE_INDEXNOW_KEY;
 
 function generateIndexNow() {
   console.log('[SEO] Starting IndexNow key verification file generation...');
   
-  if (!env.VITE_INDEXNOW_KEY) {
-    console.warn('[SEO Warning] VITE_INDEXNOW_KEY is not defined in environment. Using stable default key.');
+  if (!INDEXNOW_KEY) {
+    throw new Error('VITE_INDEXNOW_KEY is not defined in the environment. IndexNow verification file generation failed.');
   }
 
   const keyFilename = `${INDEXNOW_KEY}.txt`;
